Fix signed overflow due to implicit integer conversion

+++ lib/source/aes_encrypt.c
@@ -64,7 +64,7 @@ static inline unsigned int rotword(unsigned int a)
 	return (((a) >> 24)|((a) << 8));
 }
 
-#define subbyte(a, o)(sbox[((a) >> (o))&0xff] << (o))
+#define subbyte(a, o)((unsigned int)sbox[((a) >> (o))&0xff] << (o))
 #define subword(a)(subbyte(a, 24)|subbyte(a, 16)|subbyte(a, 8)|subbyte(a, 0))
 
 int tc_aes128_set_encrypt_key(TCAesKeySched_t s, const uint8_t *k)
@@ -83,8 +83,8 @@ int tc_aes128_set_encrypt_key(TCAesKeySched_t s, const uint8_t *k)
 	}
 
 	for (i = 0; i < Nk; ++i) {
-		s->words[i] = (k[Nb*i]<<24) | (k[Nb*i+1]<<16) |
-			      (k[Nb*i+2]<<8) | (k[Nb*i+3]);
+		s->words[i] = ((unsigned int)k[Nb*i]<<24) | ((unsigned int)k[Nb*i+1]<<16) |
+			      ((unsigned int)k[Nb*i+2]<<8) | ((unsigned int)k[Nb*i+3]);
 	}
 
 	for (; i < (Nb * (Nr + 1)); ++i) {
